requirements.txt
gunicorn==22.0.0
aiobotocore==2.15.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
asciitree==0.3.3
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
botocore==1.35.16
certifi==2024.8.30
cftime==1.6.4
click==8.1.7
fasteners==0.19
fastparquet==2024.5.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
fsspec==2024.9.0
h5netcdf==1.3.0
h5py==3.12.1
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.1.0
netCDF4==1.7.1.post2
numcodecs==0.13.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3fs==2024.9.0
scipy==1.14.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
Werkzeug==3.0.4
wrapt==1.16.0
xarray==2024.9.0
yarl==1.11.1
zarr==2.18.3
zipp==3.20.1
